requirements.txt
beautifulsoup4
tgcrypto
pyrogram
requests
pymongo[srv]
pySmartDL
python-dotenv
hachoir